Create README.md
I had to mothball it as the pandemic hit but before that was naking very good progress. The business plan and investor pitch have been well received by equity investors and we were well positioned to gain substantial funding . We've bootstrapped so far and to relaunch are lookign for a visionary coder who can contribute knowledge and time to bring the app up to date and reposition for investment in return for equity as a shareholder.